refactor(post): controller 책임 분리
